// frontend/src/app/model/influencer.model.ts
import { SocialMedia } from './social-media.model';

export class Influencer {

  id: number;
  name: string;
  lastName: string;
  gender: string;
  instagram: SocialMedia;
  twitter: SocialMedia;
  youtube: SocialMedia;

  // constructor(id: number, name: string, lastName: string, gender: string,
  // instagram: SocialMedia, twitter: SocialMedia, youtube: SocialMedia) {
  //   this.name = name;
  //   this.lastName = lastName;
  //   this.gender = gender;
  //   this.instagram = instagram;
  //   this.twitter = twitter;
  //   this.youtube = youtube;
  // }
}